completers: Use ipython logic for 'import' completions
