+++ xonsh/completers/python.py
@@ -1,9 +1,7 @@
 """Completers for Python code"""
 import re
-import sys
 import inspect
 import builtins
-import importlib
 import warnings
 import collections.abc as cabc
 from xonsh.parsers.completion_context import CompletionContext, PythonContext
@@ -281,46 +279,3 @@ def python_signature_complete(prefix, line, end, ctx, filter_func):
     args = {p + "=" for p in sig.parameters if filter_func(p, prefix)}
     return args
 
-
-@contextual_completer
-def complete_import(context: CompletionContext):
-    """
-    Completes module names and objects for "import ..." and "from ... import
-    ...".
-    """
-    if not (context.command and context.python):
-        # Imports are only possible in independent lines (not in `$()` or `@()`).
-        # This means it's python code, but also can be a command as far as the parser is concerned.
-        return None
-
-    command = context.command
-
-    if command.opening_quote:
-        # can't have a quoted import
-        return None
-
-    arg_index = command.arg_index
-    prefix = command.prefix
-    args = command.args
-
-    if arg_index == 1 and args[0].value == "from":
-        # completing module to import
-        return {"{} ".format(i) for i in complete_module(prefix)}
-    if arg_index >= 1 and args[0].value == "import":
-        # completing module to import
-        return complete_module(prefix)
-    if arg_index == 2 and args[0].value == "from":
-        return {RichCompletion("import", append_space=True)}
-    if arg_index > 2 and args[0].value == "from" and args[2].value == "import":
-        # complete thing inside a module
-        try:
-            mod = importlib.import_module(args[1].value)
-        except ImportError:
-            return set()
-        out = {i[0] for i in inspect.getmembers(mod) if i[0].startswith(prefix)}
-        return out
-    return set()
-
-
-def complete_module(prefix):
-    return {s for s in sys.modules if get_filter_function()(s, prefix)}
